+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 0.9.11
+
+* `Versions`:
+    * `Klock`: `2.6.1` -> `2.6.2`
+* `Coroutines`:
+    * `Compose`:
+        * Created :)
+        * New extensions and function:
+            * `Composition#linkWithJob`
+            * `Composition#linkWithContext`
+            * `renderComposableAndLinkToContext`
+
 ## 0.9.10
 
 * `Versions`: